Added Auth Routes (#11)
